pyproject.toml
[project]
name = "narrative-recon"
version = "0.1.0"
description = "Scrape news articles, reconstruct narratives, and analyze them using Google GenAI."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "celery>=5.5.3",
    "dateparser>=1.2.2",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.116.1",
    "flower>=2.0.1",
    "google-genai>=1.27.0",
    "lxml>=6.0.0",
    "nepali-datetime>=1.0.8.4",
    "networkx>=3.5",
    "pandas>=2.3.1",
    "psycopg>=3.2.9",
    "pytest>=8.4.1",
    "redis>=6.3.0",
    "requests>=2.32.4",
    "uvicorn>=0.35.0",
]

[tool.ruff]
line-length = 79
indent-width = 4

[tool.ruff.format]
docstring-code-format = true
docstring-code-line-length = 72
quote-style = "single"
indent-style = "space"

[dependency-groups]
dev = ["ruff>=0.12.7"]
